+++ otserv/trunk/ioaccount.h
@@ -33,9 +33,9 @@ class IOAccount {
 	/** Get a textual description of what source is used
 	  * \returns Name of the source*/
 	virtual char* getSourceDescription(){return "Account source: NULL";};
-	virtual Account loadAccount(unsigned long accno);
+	virtual Account loadAccount(uint32_t accno);
 	
-	virtual bool getPassword(unsigned long accno, const std::string& name, std::string& password);
+	virtual bool getPassword(uint32_t accno, const std::string& name, std::string& password);
 
   protected:
 	IOAccount(){};
